requirements.txt
boto3
pytest
moto
python-dotenv
